All testing is now in mesontest.py, which simplifies a lot of stuff.
